Pull up `getNext` method
